Move the helper functions up to a utils module
